Move gen-id.js
