+++ lib/index.js
@@ -2,7 +2,7 @@ const path = require('path')
 const querystring = require('querystring')
 const compiler = require('vue-template-compiler')
 const transpile = require('vue-template-es2015-compiler')
-const genId = require('./gen-id')
+const genId = require('./modules/gen-id')
 
 module.exports = function (content) {
   this.cacheable()
